backend/requirements.txt
fastapi==0.128.5
uvicorn[standard]==0.35.0
motor==3.7.1
pydantic==2.12.3
python-jose[cryptography]==3.3.0
python-multipart==0.0.9
pdfplumber==0.11.4
python-docx==1.1.2
reportlab==4.2.5
scikit-learn==1.5.2
openai==1.54.4
nltk==3.9.1
python-dotenv==1.0.1
cloudinary==1.41.0
redis==5.2.1
pytest==8.3.3
httpx==0.27.2
